# Changes   o Updated Assets for demonstration     - Basic vertex and fragment shader   o Added Renderer Command, Camera and VertexArray abstraction plus the OpenGL implementation   o Updated RendererAPI definition, Buffer, Shader abstraction plus the OpenGL implementation   o Updated Build System   o Updated Library

+++ Source/Engine/Platform/OpenGL/GLBuffer.cpp
@@ -4,7 +4,7 @@
 namespace Ultra {
 
 // Vertex Buffer
-GLVertexBuffer::GLVertexBuffer(float *vertices, uint32_t size) {
+GLVertexBuffer::GLVertexBuffer(float *vertices, size_t size) {
 	glCreateBuffers(1, &RendererID);
 	glBindBuffer(GL_ARRAY_BUFFER, RendererID);
 	glBufferData(GL_ARRAY_BUFFER, size, vertices, GL_STATIC_DRAW);
